modify: 优化message初始化
